@@ -1,5 +1,5 @@
 module Experimentation
-  class TreatmentGroupMembership < ActiveRecord::Base
+  class TreatmentGroupMembership < ApplicationRecord
     belongs_to :treatment_group
     belongs_to :patient
     belongs_to :experiment
@@ -58,14 +58,20 @@ module Experimentation
         visit_facility_block: visit_facility.block,
         visit_facility_district: visit_facility.district,
         visit_facility_state: visit_facility.state,
-        status: :visited,
-        status_updated_at: Time.current,
-        status_reason: :visit_recorded,
-        days_to_visit: days_to_visit
+        days_to_visit: days_to_visit,
+        **visit_status_fields
       )
     end
 
     private
+
+    def visit_status_fields
+      return {} if status_evicted?
+
+      {status: :visited,
+       status_reason: :visit_recorded,
+       status_updated_at: Time.current}
+    end
 
     def one_active_experiment_per_patient
       existing_memberships =
