@@ -65,7 +65,7 @@ class DHIS2Exporter
 
   def reporting_period(month_period)
     if Flipper.enabled?(:dhis2_use_ethiopian_calendar)
-      EthiopiaCalendarUtilities.gregorian_month_period_to_ethiopian(month_period).to_dhis2
+      EthiopiaCalendarUtilities.gregorian_month_period_to_ethiopian(month_period).to_s(:dhis2)
     else
       month_period.to_s(:dhis2)
     end
