@@ -149,7 +149,7 @@ class Patient < ApplicationRecord
   end
 
   def latest_mobile_number
-    phone_numbers.phone_type_mobile.last&.number_with_country_code
+    phone_numbers.phone_type_mobile.last&.localized_phone_number
   end
 
   def latest_bp_passport
