@@ -1,5 +1,8 @@
 class PatientPhoneNumber < ApplicationRecord
   include Mergeable
+  include PhoneNumberLocalizable
+
+  alias_attribute :phone_number, :number
 
   enum phone_type: {
     mobile: "mobile",
@@ -53,11 +56,5 @@ class PatientPhoneNumber < ApplicationRecord
     end
 
     exotel_phone_number_detail.update(whitelist_requested_at: time)
-  end
-
-  def number_with_country_code
-    parsed_number = Phonelib.parse(number, Rails.application.config.country[:abbreviation]).raw_national
-    default_country_code = Rails.application.config.country[:sms_country_code]
-    default_country_code + parsed_number
   end
 end
