@@ -17,7 +17,7 @@ class UserAnalyticsPresenter
   HTN_CONTROL_MONTHS_AGO = 12
   TROPHY_MILESTONES = [10, 25, 50, 100, 250, 500, 1_000, 2_000, 3_000, 4_000, 5_000]
   TROPHY_MILESTONE_INCR = 10_000
-  CACHE_VERSION = 3
+  CACHE_VERSION = 4
   EXPIRE_STATISTICS_CACHE_IN = 15.minutes
 
   def daily_stats_by_date(*stats)
@@ -372,7 +372,8 @@ class UserAnalyticsPresenter
   end
 
   def statistics_cache_key
-    "user_analytics/#{current_facility.id}/#{diabetes_enabled?}/#{CACHE_VERSION}"
+    v2 = Reports.reporting_schema_v2?
+    "user_analytics/#{current_facility.id}/dm=#{diabetes_enabled?}/v2=#{v2}/#{CACHE_VERSION}"
   end
 
   def sum_by_gender(data)
