@@ -2,21 +2,84 @@ class CohortService
   include BustCache
   CACHE_VERSION = 3
   CACHE_TTL = 7.days
+
+  attr_reader :field_prefix
   attr_reader :periods
   attr_reader :region
+  attr_reader :region_field
+  attr_reader :reporting_schema_v2
 
-  def initialize(region:, periods:)
-    @region = region
-    @periods = periods
+  COUNTS = %i[
+    cohort_controlled
+    cohort_missed_visit
+    cohort_patients
+    cohort_uncontrolled
+  ].freeze
+
+  def initialize(region:, periods:, reporting_schema_v2: Reports.reporting_schema_v2?)
+    @region = region.region
+    @periods = periods.sort.reverse # Ensure we return data with most recent cohorts first
+    @reporting_schema_v2 = reporting_schema_v2
+    @region_field = "#{@region.region_type}_region_id"
+    @quarterly = @periods.first.quarter?
+    @field_prefix = quarterly? ? "quarterly" : "monthly"
+  end
+
+  def quarterly?
+    @quarterly
+  end
+
+  def sums
+    @sums ||= COUNTS.map { |field| Arel.sql("SUM(#{field_prefix}_#{field}::int) as #{field}") }
   end
 
   def call
-    periods.each_with_object([]) do |period, arry|
-      arry << compute(period)
+    if reporting_schema_v2
+      results = v2_query(periods)
+      compute_v2(results)
+    else
+      periods.each_with_object([]) do |period, arry|
+        arry << compute(period)
+      end
     end
   end
 
   private
+
+  def compute_v2(results)
+    results.each_with_object([]) do |result, arry|
+      registration_period = if quarterly?
+        result.period.previous
+      else
+        result.period.advance(months: -2)
+      end
+      arry << {
+        controlled: result.cohort_controlled,
+        no_bp: result.cohort_missed_visit,
+        period: result.period,
+        patients_registered: registration_period.to_s,
+        registered: result.cohort_patients,
+        results_in: result.period.to_s(:cohort),
+        uncontrolled: result.cohort_uncontrolled
+      }.with_indifferent_access
+    end
+  end
+
+  def v2_query(range)
+    if quarterly?
+      range = range.map { |p| p.to_s(:quarter_string) }
+      Reports::QuarterlyFacilityState.where(facility: region.facilities, quarter_string: range)
+        .group(region_field, :quarter_string, :month_date)
+        .order("quarter_string desc")
+        .select(:month_date, region_field, sums)
+    else
+      range = periods.map { |p| p.value }
+      Reports::FacilityState.where(facility: region.facilities, month_date: range)
+        .group(region_field, :month_date)
+        .order("month_date desc")
+        .select(:month_date, region_field, sums)
+    end
+  end
 
   def compute(period)
     Rails.cache.fetch(cache_key(period), version: cache_version, expires_in: CACHE_TTL, force: bust_cache?) do
@@ -32,11 +95,12 @@ class CohortService
              registration_month: cohort_period.value.try(:month)}
       query = ControlRateCohortQuery.new(facilities: region.facilities, cohort_period: hsh)
       {
-        results_in: results_in,
-        patients_registered: cohort_period.to_s,
-        registered: query.cohort_patients.count,
         controlled: query.cohort_controlled_bps.count,
         no_bp: query.cohort_missed_visits_count,
+        patients_registered: cohort_period.to_s,
+        period: period,
+        registered: query.cohort_patients.count,
+        results_in: results_in,
         uncontrolled: query.cohort_uncontrolled_bps.count
       }.with_indifferent_access
     end
