@@ -1,2 +1,2 @@
 # encoding: UTF-8
-DataMigrate::Data.define(version: 20211115131602)
+DataMigrate::Data.define(version: 20211123060536)
